paper_scraper/requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.3
feedparser==6.0.12
greenlet==3.2.4
idna==3.10
psycopg2-binary==2.9.10
PyMuPDF==1.23.26
PyMuPDFb==1.23.22
python-dotenv==1.1.1
requests==2.32.5
sgmllib3k==1.0.0
SQLAlchemy==2.0.43
typing_extensions==4.15.0
urllib3==2.5.0
beautifulsoup4==4.14.3
